Move IntIncrementNode to supernodes package
